Basic config in full
